Add glyph_module and glyph API and reimplement with repeat modes to replace axes and mirror glyphs, static labels, label offset. Standardise arrow and axis glyphs and remove different resolutions of glyphs e.g. cylinder6, sphere_hires etc. from base library. Issue 2107.

+++ core/source/general/indexed_list_stl_private.hpp
@@ -205,7 +205,7 @@ PROTOTYPE_NUMBER_IN_LIST_FUNCTION(object_type) \
 	if (list) \
 	{ \
 		CMISS_SET(object_type) *cmiss_set = reinterpret_cast<CMISS_SET(object_type) *>(list); \
-		return cmiss_set->size(); \
+		return static_cast<int>(cmiss_set->size()); \
 	} \
 	display_message(ERROR_MESSAGE, \
 		"NUMBER_IN_LIST(" #object_type ").  Invalid argument"); \
